# Conflicts:
#	app/src/main/kotlin/com/rahim/data/reminder/YadinoBroadCastReceiver.kt

+++ core/db/src/main/java/com/rahim/yadino/db/routine/model/RoutineEntity.kt
@@ -17,6 +17,6 @@ data class RoutineEntity(
   val id: Int? = null,
   val explanation: String? = null,
   val isSample: Boolean = false,
-  val idAlarm: Long? = null,
+  val idAlarm: Int? = null,
   val timeInMillisecond: Long? = null,
 )
